backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.5
chromadb==1.0.20
openai==1.104.0
google-generativeai==0.8.5
pymupdf==1.26.4
python-multipart==0.0.20
python-dotenv==1.1.1
aiofiles==24.1.0
requests==2.32.5
pydantic==2.11.7